source/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
python-dateutil==2.8.2
PyQt5==5.15.9
selenium==4.18.1
gspread==6.0.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0

# Browser automation
webdriver-manager==4.0.1  # For ChromeDriver management

# Sub-dependencies
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.19.2
google-api-python-client==2.123.0
googleapis-common-protos==1.63.0
google-auth-httplib2==0.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
outcome==1.3.0
packaging==23.2
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pysocks==1.7.1
python-dotenv==1.0.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.2.0
wsproto==1.2.0
apscheduler==3.10.1

# Mac-specific dependencies
pyobjc-core==10.1 ; sys_platform == 'darwin'  # Required for PyQt5 on macOS
pyobjc-framework-Cocoa==10.1 ; sys_platform == 'darwin'  # Required for PyQt5 on macOS